Removed: revert all the modifications made from the open core version (hg) because of the random server bug

+++ src/ai_service/ais_actions.h
@@ -282,7 +282,7 @@ public:
 	static uint size()
 	{
 		if (!initialised()) init();
-		return (uint)Stack.size();
+		return Stack.size();
 	}
 
 	static void pop()
