+++ src/entities_game_service/egs_sheets/egs_static_game_sheet.h
@@ -185,7 +185,7 @@ public :
 	/// called when the sheet is removed
 	void removed() {}
 
-	/// called to copy from another sheet (operator= + care ptrs)
+	/// called to copy from an other sheet (operator= + care ptrs)
 	void reloadSheet(const CStaticXpStagesTable &o);
 	
 	std::vector< SStageTable > XpStagesTables;
@@ -536,7 +536,7 @@ public:
 	virtual float							getRegen(size_t index) const { return _Regen[index]; }
 	virtual SMirrorEquipment const&			getItems(size_t index) const { return _Items[index]; }
 	virtual std::string						getLootTable(uint i) const { if (i<_LootTables.size()) return _LootTables[i]; else return ""; }
-	virtual uint							getLootTableCount() const { return (uint)_LootTables.size(); }
+	virtual uint							getLootTableCount() const { return _LootTables.size(); }
 	virtual uint8							getMeleeReachValue() const { return _MeleeReachValue; }
 	
 	virtual bool							getDodgeAsDefense() const { return _DodgeAsDefense; }
@@ -647,7 +647,7 @@ public:
 	/// called when the sheet is removed
 	void removed() { }
 	
-	/// called to copy from another sheet (operator= + care ptrs)
+	/// called to copy from an other sheet (operator= + care ptrs)
 	void reloadSheet(const CStaticCreatures &o);
 	
 	/// recompute CreatureDamagePerHit according to success chance table and CreatureDamagePerHitWithoutAverageDodge
@@ -1145,7 +1145,7 @@ public :
 			}
 			else
 			{
-				uint16 size = (uint16)ChildSkills.size();
+				uint16 size = ChildSkills.size();
 				f.serial( size );
 				for( std::vector<SKILLS::ESkills>::iterator it = ChildSkills.begin(); it != ChildSkills.end(); ++it )
 				{
