+++ src/logger_service/log_storage.h
@@ -79,7 +79,7 @@ public:
 			s.serial(ContextStack);
 			s.serial(ShardId);
 			s.serialCont(ParamIndex);
-			uint32 nbList = (uint32)ListParamIndex.size();
+			uint32 nbList = ListParamIndex.size();
 			s.serial(nbList);
 			ListParamIndex.resize(nbList);
 			for (uint i=0; i<ListParamIndex.size(); ++i)
@@ -271,7 +271,7 @@ public:
 
 		dle.ContextStack = _ContextStack;
 		// set the type of the log
-		dle.LogType = (uint32)_LogDefIndex[logEntry.LogInfo.getLogName()];
+		dle.LogType = _LogDefIndex[logEntry.LogInfo.getLogName()];
 
 		// post increment if open context
 		if (dle.LogDate == 0)
@@ -286,7 +286,7 @@ public:
 			lpi.ParamType = pv.getType();
 
 			TParamsTable &pt = _ParamTables[lpi];
-			uint32 index = (uint32)pt.size();
+			uint32 index = pt.size();
 			pt.push_back(pv);
 
 			// store the index in the persistent log entry
@@ -308,7 +308,7 @@ public:
 			std::list < LGS::TParamValue >::const_iterator first(lpv.getParams().begin()), last(lpv.getParams().end());
 			for (; first != last; ++first)
 			{
-				uint32 index = (uint32)pt.size();
+				uint32 index = pt.size();
 				pt.push_back(*first);
 
 				// store the index in the persistent log entry
@@ -328,7 +328,7 @@ public:
 		// serial the log entryes
 		s.serialCont(_DiskLogEntries);
 		// serial the param tables
-		uint32 nbTable = (uint32)_ParamTables.size();
+		uint32 nbTable = _ParamTables.size();
 		if (s.isReading())
 		{
 			s.serial(nbTable);
